Creación del ControllerVendedorView
